requirements.txt
torch
transformers
pytesseract
python-dotenv
pydantic
fastapi
uvicorn
groq
streamlit